requirements.txt
Django==1.6.6
django-extensions==1.4.9
django-mptt==0.6.1
gnureadline==6.3.3
gunicorn==19.1.1
ipython==2.3.1
Pillow==2.7.0
psycopg2==2.5.3
six==1.9.0
